Move DataListMetadata to seperate file
