+++ src/Altinn.App.Core/Features/Options/AppOptionsService.cs
@@ -32,4 +32,3 @@ namespace Altinn.App.Core.Features.Options
         }
     }
 }
-
